renamed bitstampede.rb --> Bitstampede.rb
