+++ lib/bitstampede/Client.rb
